requirements.txt
Django==4.2
python-decouple==3.8
djangorestframework==3.14.0
psycopg==3.1.9
django-oauth-toolkit==2.3.0